# OEUPB-Docs/specs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OE UPB - Observatorio de Egresados",
    "description": "API RESTful para el procesamiento de egresados UPB.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/publicaciones/": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "summary": "Listar Publicaciones Autorizadas",
        "operationId": "listar_publicaciones_autorizadas_api_publicaciones__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicacionResponse"
                  },
                  "type": "array",
                  "title": "Response Listar Publicaciones Autorizadas Api Publicaciones  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Publicaciones"
        ],
        "summary": "Publicar Grafica",
        "operationId": "publicar_grafica_api_publicaciones__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicacionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionResponse"
                }
              }
            }
          },
          "403": {
            "description": "El actor no es coordinador o no tiene sede"
          },
          "422": {
            "description": "Definición no publicable o sin celdas que superen el umbral mínimo de privacidad"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/publicaciones/{publicacion_id}": {
      "delete": {
        "tags": [
          "Publicaciones"
        ],
        "summary": "Retirar Publicacion",
        "operationId": "retirar_publicacion_api_publicaciones__publicacion_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publicacion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publicacion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/publicaciones/mias": {
      "get": {
        "tags": [
          "Publicaciones"
        ],
        "summary": "Listar Publicaciones Propias",
        "operationId": "listar_publicaciones_propias_api_publicaciones_mias_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicacionResponse"
                  },
                  "type": "array",
                  "title": "Response Listar Publicaciones Propias Api Publicaciones Mias Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetGrafica": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 150,
            "title": "Label"
          },
          "data": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Data"
          },
          "backgroundColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backgroundcolor"
          },
          "borderColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bordercolor"
          }
        },
        "type": "object",
        "required": [
          "label",
          "data"
        ],
        "title": "DatasetGrafica"
      },
      "DefinicionGrafica": {
        "properties": {
          "origen": {
            "type": "string",
            "enum": [
              "reporte_general",
              "tendencias",
              "explorador"
            ],
            "title": "Origen"
          },
          "tipo_visualizacion": {
            "type": "string",
            "enum": [
              "bar",
              "line",
              "pie",
              "doughnut"
            ],
            "title": "Tipo Visualizacion"
          },
          "indicador": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Indicador"
          },
          "pregunta": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Pregunta"
          },
          "momento": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  0,
                  1,
                  5
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Momento"
          },
          "programa": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Programa"
          },
          "anio": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2200.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Anio"
          }
        },
        "type": "object",
        "required": [
          "origen",
          "tipo_visualizacion"
        ],
        "title": "DefinicionGrafica"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricasGrafica": {
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Labels"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetGrafica"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Datasets"
          }
        },
        "type": "object",
        "required": [
          "labels",
          "datasets"
        ],
        "title": "MetricasGrafica"
      },
...
      "PublicacionCreate": {
        "properties": {
          "grafica_key": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[a-z0-9_\\-]+$",
            "title": "Grafica Key"
          },
          "titulo": {
            "type": "string",
            "maxLength": 180,
            "minLength": 3,
            "title": "Titulo"
          },
          "definicion": {
            "$ref": "#/components/schemas/DefinicionGrafica"
          },
          "aprobada_privacidad": {
            "type": "boolean",
            "const": true,
            "title": "Aprobada Privacidad",
            "description": "Confirmación explícita del coordinador propietario de que revisó la privacidad de la gráfica"
          }
        },
        "type": "object",
        "required": [
          "grafica_key",
          "titulo",
          "definicion",
          "aprobada_privacidad"
        ],
        "title": "PublicacionCreate",
        "description": "Solicitud de publicación.\n\nEl cliente solo describe qué gráfica publicar. Las métricas y los programas de\naudiencia los recalcula el backend con los datos de la sede del JWT; cualquier\ncampo adicional enviado por el cliente se ignora."
      },
      "PublicacionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "grafica_key": {
            "type": "string",
            "title": "Grafica Key"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "sede_id": {
            "type": "integer",
            "title": "Sede Id"
          },
          "sede_nombre": {
            "type": "string",
            "title": "Sede Nombre"
          },
          "programas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Programas"
          },
          "permiso_requerido": {
            "type": "string",
            "title": "Permiso Requerido"
          },
          "definicion": {
            "$ref": "#/components/schemas/DefinicionGrafica"
          },
          "metricas": {
            "$ref": "#/components/schemas/MetricasGrafica"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "aprobada_privacidad": {
            "type": "boolean",
            "title": "Aprobada Privacidad"
          },
          "fecha_publicacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Publicacion"
          },
          "fecha_retiro": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Retiro"
          }
        },
        "type": "object",
        "required": [
          "id",
          "grafica_key",
          "titulo",
          "sede_id",
          "sede_nombre",
          "programas",
          "permiso_requerido",
          "definicion",
          "metricas",
          "version",
          "estado",
          "aprobada_privacidad",
          "fecha_publicacion",
          "fecha_retiro"
        ],
        "title": "PublicacionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}